requirements.txt
boto3
ask_sdk_core
ask_sdk_model
pillow